backend/requirements.txt
fastapi
uvicorn[standard]
scikit-learn
xgboost
pandas
joblib
beanie
motor
pydantic[email]
python-dotenv
